+++ sabi/src/main/java/com/sabi/sabi/dto/DiagnosticoDTO.java
@@ -40,9 +40,9 @@ public class DiagnosticoDTO {
     private Double circunferenciaCadera;
     private Double circunferenciaBrazo;
     private Double circunferenciaPierna;
-    private Integer frecuenciaCardiacaReposo; // bpm
+    private Long frecuenciaCardiacaReposo; // bpm
     private String presionArterial;           // ej. "120/80"
-    private Integer horasSueno;               // horas promedio
+    private Long horasSueno;               // horas promedio
     private String habitosAlimenticios;       // descripción libre
 
     private Boolean estado = true;
